fix(transactions): infinite loop
Currently the API to ONE Finance was changed ever so slightly to always
return a value for next, even when there are no more records to fetch.
This is a problem because it creates an infinite loop causing the
program to never finish downloading transactions, which eventually
causes a 500 error on the ONE server. Fix this by checking the value of
the next variable by decoding it to see if there is actually anything
else to fetch